moving validateName to the global validation utils module
