+++ src/validation/blueprint/update.ts
@@ -1,5 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
-import { validateName, validateFields } from './utils';
+import { validateFields } from './utils';
+import { validateName } from '../utils';
 
 export const update = async (req: Request, res: Response, next: NextFunction) => {
   const { name, fields } = req.body;
