+++ src/validation/blueprint/utils/index.ts
@@ -1,2 +1,1 @@
-export { validateName } from './validate_name';
 export { validateFields } from './validate_fields';
