+++ src/validation/utils/index.ts
@@ -3,6 +3,8 @@ export { compareType } from './compare_type';
 export { getModelInstance } from './get_model_instance';
 export { getPaginationData } from './get_pagination_data';
 export { escapeRegex } from './escape_regex';
+export { validateName } from './validate_name';
+
 export {
   ValidationError,
   ModelTypes,
